requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.4
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.0
django-allauth==0.52.0
dj-rest-auth==4.0.0
django_renderpdf==4.0.0